Corrige el acceso al ultimo estado al realizar un cambio de estado. Mueve la ruta de camas al modulo rup.
