+++ modules/rup/schemas/camaEstado.ts
@@ -1,6 +1,6 @@
-import { pacienteSchema } from './../../mpi/schemas/paciente';
+import { pacienteSchema } from '../../../core/mpi/schemas/paciente';
 import * as mongoose from 'mongoose';
-import { SnomedConcept } from '../../../modules/rup/schemas/snomed-concept';
+import { SnomedConcept, ISnomedConcept } from './snomed-concept';
 import { AuditPlugin } from '@andes/mongoose-plugin-audit';
 
 export let schema = new mongoose.Schema({
